Split into local and remote calls
