+++ src/Features/Core/Portable/Extensions/ExtensionMessageHandlerService.cs
@@ -15,7 +15,6 @@ using System.Threading.Tasks;
 using Microsoft.CodeAnalysis.ErrorReporting;
 using Microsoft.CodeAnalysis.Host;
 using Microsoft.CodeAnalysis.PooledObjects;
-using Microsoft.CodeAnalysis.Remote;
 using Microsoft.CodeAnalysis.Threading;
 using Roslyn.Utilities;
 
@@ -66,55 +65,11 @@ internal sealed partial class ExtensionMessageHandlerServiceFactory
                 ?? throw new InvalidOperationException($"Unable to get the directory name for {assemblyFilePath}.");
         }
 
-        private async ValueTask<TResult> ExecuteInRemoteOrCurrentProcessAsync<TResult>(
-            Solution? solution,
-            Func<CancellationToken, ValueTask<TResult>> executeInProcessAsync,
-            Func<IRemoteExtensionMessageHandlerService, Checksum?, CancellationToken, ValueTask<TResult>> executeOutOfProcessAsync,
-            CancellationToken cancellationToken)
-        {
-            var client = await RemoteHostClient.TryGetClientAsync(this.SolutionServices, cancellationToken).ConfigureAwait(false);
-            if (client is null)
-                return await executeInProcessAsync(cancellationToken).ConfigureAwait(false);
-
-            if (solution is null)
-            {
-                var result = await client.TryInvokeAsync<IRemoteExtensionMessageHandlerService, TResult>(
-                    (remoteService, cancellationToken) => executeOutOfProcessAsync(remoteService, null, cancellationToken),
-                    cancellationToken).ConfigureAwait(false);
-
-                // If the remote call succeeded, this will have a valid valid in it and can be returned.  If it did not
-                // succeed then we will have already shown the user an error message stating there was an issue making the call,
-                // and it's fine for this to throw again, unwinding the stack up to the caller.
-                return result.Value;
-            }
-            else
-            {
-                var result = await client.TryInvokeAsync<IRemoteExtensionMessageHandlerService, TResult>(
-                    solution,
-                    (remoteService, checksum, cancellationToken) => executeOutOfProcessAsync(remoteService, checksum, cancellationToken),
-                    cancellationToken).ConfigureAwait(false);
-                return result.Value;
-            }
-        }
-
         private void ClearCachedHandlers_WhileUnderLock()
         {
             Contract.ThrowIfTrue(!Monitor.IsEntered(_gate));
             _cachedWorkspaceHandlers_useOnlyUnderLock.Clear();
             _cachedDocumentHandlers_useOnlyUnderLock.Clear();
-        }
-
-        #region RegisterExtension
-
-        public async ValueTask RegisterExtensionAsync(
-            string assemblyFilePath,
-            CancellationToken cancellationToken)
-        {
-            await ExecuteInRemoteOrCurrentProcessAsync(
-                solution: null,
-                _ => RegisterExtensionInCurrentProcessAsync(assemblyFilePath),
-                (remoteService, _, cancellationToken) => remoteService.RegisterExtensionAsync(assemblyFilePath, cancellationToken),
-                cancellationToken).ConfigureAwait(false);
         }
 
         public ValueTask<VoidResult> RegisterExtensionInCurrentProcessAsync(string assemblyFilePath)
@@ -133,21 +88,6 @@ internal sealed partial class ExtensionMessageHandlerServiceFactory
                 ClearCachedHandlers_WhileUnderLock();
                 return default;
             }
-        }
-
-        #endregion
-
-        #region UnregisterExtension
-
-        public async ValueTask UnregisterExtensionAsync(
-            string assemblyFilePath,
-            CancellationToken cancellationToken)
-        {
-            await ExecuteInRemoteOrCurrentProcessAsync(
-                solution: null,
-                _ => UnregisterExtensionInCurrentProcessAsync(assemblyFilePath),
-                (remoteService, _, cancellationToken) => remoteService.UnregisterExtensionAsync(assemblyFilePath, cancellationToken),
-                cancellationToken).ConfigureAwait(false);
         }
 
         private ValueTask<VoidResult> UnregisterExtensionInCurrentProcessAsync(string assemblyFilePath)
@@ -172,21 +112,6 @@ internal sealed partial class ExtensionMessageHandlerServiceFactory
             }
         }
 
-        #endregion
-
-        #region GetExtensionMessageNames
-
-        public async ValueTask<GetExtensionMessageNamesResponse> GetExtensionMessageNamesAsync(
-            string assemblyFilePath,
-            CancellationToken cancellationToken)
-        {
-            return await ExecuteInRemoteOrCurrentProcessAsync(
-                solution: null,
-                cancellationToken => GetExtensionMessageNamesInCurrentProcessAsync(assemblyFilePath, cancellationToken),
-                (remoteService, _, cancellationToken) => remoteService.GetExtensionMessageNamesAsync(assemblyFilePath, cancellationToken),
-                cancellationToken).ConfigureAwait(false);
-        }
-
         public async ValueTask<GetExtensionMessageNamesResponse> GetExtensionMessageNamesInCurrentProcessAsync(
             string assemblyFilePath,
             CancellationToken cancellationToken)
@@ -207,19 +132,6 @@ internal sealed partial class ExtensionMessageHandlerServiceFactory
                 [.. assemblyHandlers.DocumentMessageHandlers.Keys]);
         }
 
-        #endregion
-
-        #region Reset
-
-        public async ValueTask ResetAsync(CancellationToken cancellationToken)
-        {
-            await ExecuteInRemoteOrCurrentProcessAsync(
-                solution: null,
-                _ => ResetInCurrentProcessAsync(),
-                (remoteService, _, cancellationToken) => remoteService.ResetAsync(cancellationToken),
-                cancellationToken).ConfigureAwait(false);
-        }
-
         private ValueTask<VoidResult> ResetInCurrentProcessAsync()
         {
             lock (_gate)
@@ -228,30 +140,6 @@ internal sealed partial class ExtensionMessageHandlerServiceFactory
                 ClearCachedHandlers_WhileUnderLock();
                 return default;
             }
-        }
-
-        #endregion
-
-        #region HandleExtensionMessage
-
-        public async ValueTask<string> HandleExtensionWorkspaceMessageAsync(Solution solution, string messageName, string jsonMessage, CancellationToken cancellationToken)
-        {
-            return await ExecuteInRemoteOrCurrentProcessAsync(
-                solution,
-                cancellationToken => HandleExtensionMessageInCurrentProcessAsync(
-                    executeArgument: solution, isSolution: true, messageName, jsonMessage, _cachedWorkspaceHandlers_useOnlyUnderLock, cancellationToken),
-                (remoteService, checksum, cancellationToken) => remoteService.HandleExtensionWorkspaceMessageAsync(checksum!.Value, messageName, jsonMessage, cancellationToken),
-                cancellationToken).ConfigureAwait(false);
-        }
-
-        public async ValueTask<string> HandleExtensionDocumentMessageAsync(Document document, string messageName, string jsonMessage, CancellationToken cancellationToken)
-        {
-            return await ExecuteInRemoteOrCurrentProcessAsync(
-                document.Project.Solution,
-                cancellationToken => HandleExtensionMessageInCurrentProcessAsync(
-                    executeArgument: document, isSolution: false, messageName, jsonMessage, _cachedDocumentHandlers_useOnlyUnderLock, cancellationToken),
-                (remoteService, checksum, cancellationToken) => remoteService.HandleExtensionDocumentMessageAsync(checksum!.Value, messageName, jsonMessage, document.Id, cancellationToken),
-                cancellationToken).ConfigureAwait(false);
         }
 
         private async ValueTask<string> HandleExtensionMessageInCurrentProcessAsync<TArgument>(
@@ -326,8 +214,6 @@ internal sealed partial class ExtensionMessageHandlerServiceFactory
 
             return result.ToImmutable();
         }
-
-        #endregion
     }
 }
 #endif
